spectrum: precompute Hann window
Computing the factor applied to each sample is the most expensive
operation in the whole process function, by a large margin. Avoid that
by precomputing the 8192 factors.

An memory size optimization could be to compute half of it, as it is
symmetrical. It would make the code more complex for small gains; we
chase the low hanging fruits here.

Here is a before/after of performances, acquired by running pw-profiler
during 15s, with a 512 samples quantum:

                 Before       After
process calls      1542        1563
average (µs)        248.1        86.2
stdev                34.6        15.8
min                  80          26
5th perc            177          58
25th perc           233          80
median              255          85
75th perc           262          93
95th perc           287         111
max                 318         208

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>