spectrum: disable ::process() if window is hidden for >200ms
We create a timeout source. Its ready time is 200ms in the future. Every
time we draw a spectrum, we move update the timeout to now+200ms. If we
reach the event, it means no frame has been drawn for 200ms. The window
is probably hidden somewhere and we can bypass Spectrum::process().

200ms is nice because it is above 8192/48000 == ~170ms, ie the longest
quantum at the default rate.

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>